Chore: Restructured Cards
