+++ src/Components/Cards/index.js
@@ -1,0 +1,23 @@
+import CommentCard from "./Comment Cards/CommentCard";
+import HomepageLoadingCard from "./HomePage Cards/HomepageLoadingCard";
+import NotificationCard from "./Notification Cards/NotificationCard";
+import NotificationLoadingCard from "./Notification Cards/NotificationLoadingCard";
+import PostCard from "./Post Cards/PostCard";
+import PostLoaderCard from "./Post Cards/PostLoaderCard";
+import ProfileCard from "./Profile Cards/ProfileCard";
+import UpdateProfileCard from "./Profile Cards/UpdateProfileCard";
+import UpdateWallColor from "./Profile Cards/UpdateWallColor";
+import SearchCard from "./Search Card/SearchCard";
+
+export {
+  CommentCard,
+  HomepageLoadingCard,
+  NotificationCard,
+  NotificationLoadingCard,
+  PostCard,
+  PostLoaderCard,
+  ProfileCard,
+  UpdateProfileCard,
+  UpdateWallColor,
+  SearchCard,
+};
