+++ src/Components/Pages/Homepage.jsx
@@ -1,6 +1,6 @@
 import TopNav from "../Navbars/Top Navs/TopNav";
 import BottomNav from "../Navbars/Bottom Navs/BottomNav";
-import HomepageLoadingCard from "../Cards/HomepageLoadingCard";
+import { HomepageLoadingCard } from "../Cards";
 import Feed from "../Feed/Feed";
 import CreatePost from "../Posts/CreatePost";
 import styles from "./hompage.module.css";
